DP, recursion + backtracking restructure
