+++ Dynamic Programming, Recursion, & Backtracking/BuyAndSellOnce/LinearSpace.java
@@ -1,4 +1,4 @@
-class LinearSpace {
+class Solution {
   public int maxProfit(int[] prices) {
     if (prices.length == 0) {
       return 0;
@@ -33,25 +33,6 @@ class LinearSpace {
 
       opt[i] = Math.max(0, opt[i - 1] + profitDelta);
       globalMax = Math.max(globalMax, opt[i]);
-    }
-
-    return globalMax;
-  }
-}
-
-class ConstantSpace {
-  public int maxProfit(int[] prices) {
-    if (prices.length == 0) {
-      return 0;
-    }
-
-    // Only 1 thing matters: lowest price seen so far, 'globalMin'
-    int globalMax = 0;
-    int globalMin = prices[0];
-
-    for (int i = 1; i < prices.length; i++) {
-      globalMin = Math.min(globalMin, prices[i]);
-      globalMax = Math.max(globalMax, prices[i] - globalMin);
     }
 
     return globalMax;
