+++ Dynamic Programming, Recursion, & Backtracking/DNASequenceAlignment/DNASequenceAlignment.java
@@ -1,72 +1,8 @@
-import java.util.*;
-
-public class DNASequenceAlignment {
+public class Solution {
   private static final int GAP_COST = 1;
   private static final Map<String, Integer> alignmentCosts = new HashMap<>();
 
-  public static void main(String args[]) {
-    String s1 = "GACGTTA";
-    String s2 = "GAACGCTA";
-
-    /*
-      Example:
-
-      g = gap
-      m = mismatch
-
-      s1 => GA-CGTTA
-      s2 => GAACGCTA
-              ^  ^
-              g  m
-      
-            0123456
-      s1 => GACGTTA
-
-            01234567
-      s2 => GAACGCTA
-
-      tuple of (i, j)
-      matching = { (0, 0), (1, 1), (2, 3), (3, 4), (4, 5), (5, 6), (6, 7) }
-                                  ^                   ^
-                        gap (s2's A unmatched)  mismatch (w/ real cost)
-    */
-
-    initializeAlignmentCosts(alignmentCosts);
-    System.out.println("Min cost to align is: " + alignmentMinCost(s1, s2));
-
-    /*
-      Trace back the table:
-          G A A C G C T A
-        0 1 2 3 4 5 6 7 8
-      G 1 0 1 2 3 4 5 6 7
-      A 2 1 0 1 2 3 4 5 6
-      C 3 2 1 1 1 2 3 4 5
-      G 4 3 2 2 2 1 2 3 4
-      T 5 4 3 3 3 2 3 2 3
-      T 6 5 4 4 4 3 4 3 4
-      A 7 6 5 4 5 4 4 4 3
-
-      1.) Add (6, 7) to the matching
-      2.) Don't match index 5 of s1 (gap in s2)
-      3.) Add (4, 6) to the matching
-      4.) Don't match index 5 of s2 (gap in s1)
-      5.) Add (3, 4) to the matching
-      6.) Add (2, 3) to the matching (we could have not matched index 3 of s2, leaving a gap in s1)
-      7.) Don't match index 2 of s2 (gap in s1)
-      8.) Add (1, 1) to the matching
-      9.) Add (0, 0) to the matching
-
-      matching = { (0, 0), (1, 1), (2, 3), (3, 4), (4, 6), (6, 7) }
-
-      Finally looks like this:
-      s1 => GA-CG-TTA
-      s2 => GAACGCT-A
-
-      Total Cost is 3 (we see that this is "gap heavy" to avoid mismatch costs)
-    */
-  }
-
-  private static int alignmentMinCost(String s1, String s2) {
+  private int alignmentMinCost(String s1, String s2) {
     int[][] opt = new int[s1.length() + 1][s2.length() + 1];
 
     // Base case #1 - all gaps (s1 is on the rows)
@@ -95,12 +31,10 @@ public class DNASequenceAlignment {
       }
     }
 
-    printTable(opt);
-
     return opt[s1.length()][s2.length()];
   }
 
-  private static void initializeAlignmentCosts(Map<String, Integer> alignmentCosts) {
+  private void initializeAlignmentCosts(Map<String, Integer> alignmentCosts) {
     alignmentCosts.put("AA", 0); // same symbol
     alignmentCosts.put("CC", 0); // same symbol
     alignmentCosts.put("GG", 0); // same symbol
@@ -129,13 +63,58 @@ public class DNASequenceAlignment {
     alignmentCosts.put("GT", 1);
     alignmentCosts.put("TG", 1);
   }
-
-  private static void printTable(int[][] opt) {
-    for (int row = 0; row < opt.length; row++) {
-      for (int col = 0; col < opt[0].length; col++) {
-        System.out.print(opt[row][col]);
-      }
-      System.out.println();
-    }
-  }
 }
+
+/*
+  Example:
+
+  g = gap
+  m = mismatch
+
+  s1 => GA-CGTTA
+  s2 => GAACGCTA
+          ^  ^
+          g  m
+  
+        0123456
+  s1 => GACGTTA
+
+        01234567
+  s2 => GAACGCTA
+
+  tuple of (i, j)
+  matching = { (0, 0), (1, 1), (2, 3), (3, 4), (4, 5), (5, 6), (6, 7) }
+                              ^                   ^
+                    gap (s2's A unmatched)  mismatch (w/ real cost)
+*/
+
+/*
+  Trace back the table:
+      G A A C G C T A
+    0 1 2 3 4 5 6 7 8
+  G 1 0 1 2 3 4 5 6 7
+  A 2 1 0 1 2 3 4 5 6
+  C 3 2 1 1 1 2 3 4 5
+  G 4 3 2 2 2 1 2 3 4
+  T 5 4 3 3 3 2 3 2 3
+  T 6 5 4 4 4 3 4 3 4
+  A 7 6 5 4 5 4 4 4 3
+
+  1.) Add (6, 7) to the matching
+  2.) Don't match index 5 of s1 (gap in s2)
+  3.) Add (4, 6) to the matching
+  4.) Don't match index 5 of s2 (gap in s1)
+  5.) Add (3, 4) to the matching
+  6.) Add (2, 3) to the matching (we could have not matched index 3 of s2, leaving a gap in s1)
+  7.) Don't match index 2 of s2 (gap in s1)
+  8.) Add (1, 1) to the matching
+  9.) Add (0, 0) to the matching
+
+  matching = { (0, 0), (1, 1), (2, 3), (3, 4), (4, 6), (6, 7) }
+
+  Finally looks like this:
+  s1 => GA-CG-TTA
+  s2 => GAACGCT-A
+
+  Total Cost is 3 (we see that this is "gap heavy" to avoid mismatch costs)
+*/
